Release ready. Collision bug still exists
This is the final release version within the scope of the thesis. Both
Uploading/Downloading works as intended, however when trying to send &
receive at the same time, there's a collision which makes the firmware
enter a locked state, which can be solved by resetting the system manually
with the reset-button on the shield.

This project continues through a summer job and the next release will be
focused on a more robust design with requests/responses.