Ready to be released. Cleaned up a bit code
This is a stable version of uploading & downloading objects. There's also
a response message for each request message. This is handled in the read()
functionality. I think that Phils workaround-FWpatch to the GW caused me
to not being able to download objects using Zorkmon. It gives me the -1004
error code. However the function should be working just fine if the GW FW
is fixed. The FOTA function is in place and has been tested with a
MAGNETIC credential only.